+++ src/main/java/umc/catchy/domain/member/dao/MemberRepository.java
@@ -1,14 +1,35 @@
 package umc.catchy.domain.member.dao;
 
 import java.util.Optional;
+
+import feign.Param;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Modifying;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.stereotype.Repository;
 import umc.catchy.domain.member.domain.Member;
 
+@Repository
 public interface MemberRepository extends JpaRepository<Member, Long> {
 
     Optional<Member> findByEmail(String email);
     Optional<Member> findByProviderId(String providerId);
     Optional<Member> findByNickname(String nickname);
     Optional<Member> findByEmailAndProviderId(String email, String providerId);
-    Optional<Member> findByAccessToken(String accessToken);
+
+    /* 멤버 삭제 시 연관 엔티티 삭제 */
+    @Modifying
+    @Query(value = """
+            DELETE FROM member_style WHERE member_id = :memberId;
+                    DELETE FROM member_place_vote WHERE member_id = :memberId;
+                    DELETE FROM member_category_vote WHERE member_id = :memberId;
+                    DELETE FROM member_active_time WHERE member_id = :memberId;
+                    DELETE FROM member_location WHERE member_id = :memberId;
+                    DELETE FROM member_group WHERE member_id = :memberId;
+                    DELETE FROM member_course WHERE member_id = :memberId;
+                    DELETE FROM member_category WHERE member_id = :memberId;
+                    DELETE FROM place_visit WHERE member_id = :memberId;
+                    DELETE FROM place_like WHERE member_id = :memberId;
+            """, nativeQuery = true)
+    void deleteAllRelatedEntities(@Param("memberId") Long memberId);
 }
